# src/wwwroot/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hein.RulesEngine.API",
        "version": "POC",
        "description": "Proof of Concept Rules Engine made by [brandonhein](https://github.com/brandonhein). Back to Admin [UI](/Prod)"
    },
    "externalDocs": {
        "description": "GitHub Repository",
        "url": "https://github.com/brandonhein/Hein.RulesEngine-v3"
    },
    "basePath": "/Prod",
    "paths": {
        "/Execute": {
            "post": {
                "tags": ["Execute"],
                "summary": "Execute a rule set",
                "operationId": "Index",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "request",
                    "in": "body",
                    "required": false,
                    "schema": {
                        "$ref": "#/definitions/ExecuteRuleModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ExecuteRuleModel"
                        }
                    }
                }
            }
        },
        "/Convert": {
            "post": {
                "tags": ["Convert"],
                "summary": "Convert Admin Code to Rules Engine Code",
                "operationId": "Index",
                "consumes": ["text/plain"],
                "produces": ["text/plain"],
                "parameters": [{
                    "name": "code",
                    "in": "body",
                    "required": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Rule": {
            "get": {
                "tags": ["RuleEntity"],
                "summary": "Get HATEOAS links for all Rules",
                "operationId": "Get_RuleEntity_1",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RuleEntityResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["RuleEntity"],
                "summary": "Create a New Rule Entity",
                "operationId": "Post_RuleEntity_1",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "entity",
                    "in": "body",
                    "required": false,
                    "schema": {
                        "$ref": "#/definitions/RuleEntityModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Rule/{ruleName}": {
            "get": {
                "tags": ["RuleEntity"],
                "summary": "Get Rule Entity",
                "operationId": "Get_RuleEntity_2",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "ruleName",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/RuleEntityModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ExecuteRuleModel": {
            "type": "object",
            "properties": {
                "rule": {
                    "type": "string"
                },
                "values": {
                    "type": "object",
                    "properties": {
                        "additionalProp1": {
                            "type": "string"
                        },
                        "additionalProp2": {
                            "type": "number"
                        },
                        "additionalProp3": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "RuleEntityModel": {
            "type": "object",
            "properties": {
                "RuleSet": {
                    "type": "string"
                },
                "AdminCode": {
                    "type": "string"
                },
                "EngineCode": {
                    "type": "string"
                }
            }
        },
        "RuleEntityResponse": {
            "type": "object",
            "properties": {
                "ruleName": {
                    "type": "string"
                },
                "_links": {
                    "type": "object",
                    "properties": {
                        "href": {
                            "type": "string"
                        },
                        "rel": {
                            "type": "string"
                        },
                        "type": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}